# api/openapi.yaml
openapi: 3.1.0
info:
  title: Synthesis server
  version: "1.0.0"
  summary: The application service and the relay boundary of the Synthesis server.
  description: |
    The REST surface of `specifications/server/SAS-server-application-service.md`,
    the relay introspection routes of
    `specifications/server/SRB-server-relay-boundary.md`, the project-discovery
    route of `specifications/server/RSN-remote-session.md`, and the three
    WebSocket upgrades of `specifications/server/WSK-websocket.md`.

    Every route and every WebSocket upgrade but `GET /v1/health` needs the header
    `Authorization: Bearer <token>`, carrying `SYNTHESIS_SERVER_TOKEN`. On the
    application routes the token authenticates the persisted administrator user
    (SAS-FR-LFCA, SAS-FR-XRPD), which bypasses every access-control check. On the
    relay it authenticates transport access alone (RSN-FR-TWQJ): the handle
    check, the worker-ownership check, and the end-to-end client authentication
    stand beside it.

    The permission each project-scoped operation needs is named in the
    description of that operation. The relay evaluates no permission.

    The frames the three WebSocket endpoints carry are not described here: a
    WebSocket protocol is outside what OpenAPI describes, and
    `specifications/server/WSK-websocket.md` holds the frame tables.
servers:
  - url: http://localhost:8080
    description: The service as it binds by default.
security:
  - bearerAuth: []
...
paths:
...
  /v1/relay/projects:
    get:
      tags: [Relay]
      summary: The projects one worker exposes to one registered client.
      description: |
        RSN-FR-MHDV. The request names a registered client with its opaque
        handle and the worker the handle is bound to. The answer holds the
        worker, its selected project, and its exposed projects, and holds no
        key, no handle, no application payload, no user, and no ownership data
        (RSN-FR-QLEC). It reports the announcement the worker route holds when
        the request is served; the relay caches nothing beyond it (RSN-FR-GXWB).
      parameters:
        - name: X-Synthesis-Client-Handle
          in: header
          required: true
          description: The opaque handle the relay assigned at registration.
          schema: { type: string }
        - name: X-Synthesis-Instance-Id
          in: header
          required: true
          description: The worker the handle is bound to.
          schema: { type: string }
      responses:
        '200':
          description: The exposed projects of that worker.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelayProjects'
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403':
          description: |
            The handle was revoked, expired, or reset (RSN-FR-TFJU). The error
            code is `handle_revoked`, `handle_expired`, or `handle_reset`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: |
            No route holds the handle, the handle is bound to another worker, or
            the worker holds no active route (RSN-FR-TFJU, RSN-FR-ZDPA). The
            error code is `handle_not_found` or `worker_not_found`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The static token of V1 (SAS-FR-LFCA), read from `SYNTHESIS_SERVER_TOKEN`.
        On the application routes it authenticates the persisted administrator
        user; on the relay routes and the WebSocket upgrades it authenticates
        transport access alone (SRB-FR-JEBM). The service compares it in constant
        time (SAS-FR-BZUK), and it reaches no answer, no log record, and no
        persisted record (SAS-FR-PMRB).
...
  responses:
...
    Unauthenticated:
      description: The request carries no accepted credential. The code is unauthenticated.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          description: The scheme the route accepts.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - unauthenticated
                - forbidden
                - not_found
                - invalid_field
                - invalid_role
                - invalid_permission
                - duplicate_id
                - duplicate_membership
                - duplicate_grant
                - owner_protected
                - referenced
                - stale_revision
                - conversation_locked
                - invitation_not_pending
                - internal
            message: { type: string }
...
    ExposedProject:
      type: object
      required: [project_id, display_name]
      properties:
        project_id:
          type: string
          description: The opaque identifier the relay routes by.
        display_name:
          type: string
          description: Presentation only, and never a routing key.
...
    RelayProjects:
      type: object
      required: [instance_id, selected_project_id, projects]
      additionalProperties: false
      properties:
        instance_id: { type: string }
        selected_project_id: { type: [string, 'null'] }
        projects:
          type: array
          items: { $ref: '#/components/schemas/ExposedProject' }